+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 ## Next Version
 
+## 4.3.0
+
 ### Added
 - Added ability to set nested template options from the command line using dot syntax eg `--option "typeAliases.ID: String"` #189
 - Added a customizable `jsonEncoder` on APIClient #172 #203
@@ -16,6 +18,8 @@
 - Add missing custom model protocol name #191
 - Fixed missing customization of JSONEncoder instance to encode request's body #147
 - Fixed string uploads #161
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/4.2.0...4.3.0)
 
 ## 4.2.0